fix points history precision

+++ packages/wallets/src/modules/assets/points/queries/get-points-asset-transactions-query-options.ts
@@ -29,7 +29,7 @@ export function getPointsAssetTransactionsQueryOptions(
         type,
         results: [
           {
-            amount: parseInt(amount),
+            amount: parseFloat(amount),
             asset: "POINTS",
           },
         ],
